+++ src/pii_data/__init__.py
@@ -1,2 +1,2 @@
 # Package version
-VERSION = "0.4.0"
+VERSION = "0.5.0"