+++ src/pii_data/types/piicollection/__init__.py
@@ -1,3 +1,3 @@
-from .collection import PiiDetector, PiiCollection
-from .loader import PiiCollectionLoader
-from .chunk import PiiChunkIterator
+from .collection import PiiDetector, PiiCollection    # noqa: F401
+from .loader import PiiCollectionLoader               # noqa: F401
+from .chunk import PiiChunkIterator                   # noqa: F401